navbar update
The menu item bug is still not solved.
Found a way around it by manually increasing the width with viewport width unit (85vw)
(menu item width works perfectly in an independent html file; there is some conflict with the main css file thats is resulting in the reduced width of the menu item.
It needs to be solved by isolating and testing the header stylesheet with different global properties from main css file)